pyproject.toml
[tool.poetry]
name = "hcoop-meetbot"
version = "0.1.0"
description = "Plugin for Limnoria to help run IRC meetings"
authors = ["Kenneth J. Pronovici <pronovic@ieee.org>"]
license = "Apache-2.0"
readme = "PyPI.md"
homepage = "https://pypi.org/project/hcoop-meetbot/"
include = [ "Changelog", "LICENSE", "CREDITS", "README.md", "docs", "tests" ]
exclude = [ "HcoopMeetbot/test.py" ] # annoyingly, bot tests must live in the source tree
packages = [ { include = "HcoopMeetbot", from = "src" },
             { include = "hcoopmeetbotlogic", from = "src" } ]
classifiers=[
   "Programming Language :: Python :: 3",
   "License :: OSI Approved :: Apache Software License",
   "Operating System :: OS Independent",
   "Environment :: Console",
   "Intended Audience :: System Administrators",
   "Natural Language :: English",
   "Topic :: Utilities",
   "Topic :: Communications :: Chat :: Internet Relay Chat"
]

[tool.poetry.plugins."limnoria.plugins"]
"HcoopMeetbot" = "HcoopMeetbot:plugin"

[tool.poetry.dependencies]
python = ">=3.7,<4"
limnoria = "^2021.01.15"
attrs = "^20.1.0"
pytz = "^2021.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
coverage = "^5.1"
tox = "^3.15.0"
pylint = "^2.6.0"
sphinx = "^3.4.3"
toml = "^0.10.1"
pre-commit = "^2.4.0"
black = "^20.8b1"
mypy = "^0.790"
isort = "^5.0.0"
sphinx-autoapi = "^1.3.0"
safety = "^1.10.3"
coveralls = "^3.0.0"

[tool.black]
line-length = 132
target-version = ['py37', 'py38']
include = '(src\/scripts\/.*$|\.pyi?$)'
exclude = '''
/(
    \.git
  | __pycache__
  | \.tox
  | \.venv
  | build
  | dist
  | docs
  | notes
)/
'''

[tool.isort]
profile = "black"
line_length = 132
skip_glob = [ "docs", "notes" ]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
